Implemented security measures for securing backend API routes.
Implemented authentication measures to secure backend API routes, ensuring that only authorized users have access to various workout endpoints.

+++ frontend/src/components/WorkoutDetails.js
@@ -1,4 +1,5 @@
 import { useWorkoutsContext } from "../hooks/useWorkoutsContext";
+import { useAuthContext } from '../hooks/useAuthContext';
 import { BsTrash } from 'react-icons/bs';
 
 //date fns
@@ -7,11 +8,19 @@ import formatDistanceToNow from 'date-fns/formatDistanceToNow'
 const WorkoutDetails = ({ workout }) => {
 
     const { dispatch } = useWorkoutsContext()
+    const { user } = useAuthContext()
 
 
     const handleClick = async () => {
+        if (!user) {
+            return
+        }
         const response = await fetch('/api/workouts/' + workout._id, {
-            method: 'DELETE'
+            method: 'DELETE',
+            headers: {
+                'Authorization': `Bearer ${user.token}`
+            }
+
         })
         const json = await response.json()
